% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_dupes.R
\name{all_dupes}
\alias{all_dupes}
\title{Find all duplicate rows}
\usage{
all_dupes(data, x)
}
\arguments{
\item{data}{A data frame or tibble}

\item{x}{The unquoted column name to check for duplicate values}
}
\value{
A tibble similar to data
}
\description{
Identifies the duplicate values in a column of a data frame (or tibble) and
returns a data frame (tibble) with just the duplicate rows.

I actually prefer \code{janitor::get_dupes()} but I am including this version
for posterity.
}
\examples{
library(dplyr)
library(tibble)
vec <- tibble::tribble(
  ~ID,      ~OS,         ~Date,
  "userA",    "Win",  "12/15/2015",
  "userB",    "OSX",    "1/1/2016",
  "userA",    "Win",  "12/15/2017",
  "userC",  "Win64",    "4/9/2018"
)

all_dupes(data = vec, x = ID)
vec \%>\%
  all_dupes(ID)
}
\references{
\url{https://stat.ethz.ch/pipermail/r-help/2011-October/291383.html}

\url{https://stackoverflow.com/questions/7854433/finding-all-duplicate-rows-including-elements-with-smaller-subscripts}
}
