% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_numeric_factor.R
\name{as_numeric_factor}
\alias{as_numeric_factor}
\title{Convert a factor to integer/numeric without loss of information}
\usage{
as_numeric_factor(x)
}
\arguments{
\item{x}{an object, for example a factor.}
}
\value{
A numeric object
}
\description{
To transform a factor \code{f} to approximately its original numeric values,
\code{as.numeric(levels(f))[f]} is recommended and slightly more efficient than
\code{as.numeric(as.character(f))}.
}
\examples{
library(dplyr)
library(tibble)

df <- tibble::tibble(
  foo = factor(sample(x = c(4, 5, 6),  size = 50, replace = TRUE),
               levels = c(6, 5, 4))
)

df

df \%>\%
  mutate(foo_base = as.numeric(foo),
         foo_lamisc = as_numeric_factor(foo),
         foo_alt = as.numeric(as.character(foo)))
}
\references{
https://stackoverflow.com/questions/3418128/how-to-convert-a-factor-to-integer-numeric-without-loss-of-information
https://cran.r-project.org/doc/FAQ/R-FAQ.html#How-do-I-convert-factors-to-numeric_003f
}
