% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_dataset.R
\docType{data}
\name{auto}
\alias{auto}
\title{Auto data set}
\format{
A data frame with 74 rows and 12 variables:
\describe{
\item{make}{MAke and model}
\item{price}{{Price}}
\item{mpg}{Miles per gallon}
\item{rep78}{Repair record (1978)}
\item{headroom}{Headroom (in.)}
\item{trunk}{Trunk space (cu. ft.)}
\item{weight}{Weight (lbs.)}
\item{length}{Length (in.)}
\item{turn}{Turn circle (ft.)}
\item{displacement}{Displacement (cu. in.)}
\item{gear_ratio}{Gear ratio}
\item{foreign}{Car type}
...
}
}
\source{
Stata
}
\usage{
auto
}
\description{
Data on various automobiles
}
\keyword{datasets}
