% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/front_and_back.R
\name{back}
\alias{back}
\alias{front}
\title{Move selected columns to front and back}
\usage{
back(data, ...)

front(data, ...)
}
\arguments{
\item{data}{A data frame or tibble}

\item{...}{Column(s) in the data frame to be moved}
}
\value{
A data frame
}
\description{
Cool functions taken from @roman_francois Twitter post.
\url{https://gist.github.com/romainfrancois/5363c96646bb164f77fac2610d23b126}
}
\details{
Given a data frame or tibble, these functions will move selected columns to
the front or to the back of the rest of the columns. Really a wrapper for
\code{dplyr::select()} and \code{dplyr::everything()} but making cool use of
tidyeval.
}
\examples{
library(dplyr)
back(iris, Species) \%>\% head
back(iris, Species, starts_with("Petal")) \%>\% head
front(iris, Species) \%>\% head
}
