% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom_ci.R
\name{binom_ci}
\alias{binom_ci}
\title{Compute Binomial Confidence Intervals}
\usage{
binom_ci(
  x,
  n,
  conf_level = 0.95,
  sides = c("two.sided", "left", "right"),
  method = c("wilson", "wald", "waldcc", "agresti-coull", "jeffreys", "modified wilson",
    "wilsoncc", "modified jeffreys", "clopper-pearson", "arcsine", "logit", "witting",
    "pratt", "midp", "lik", "blaker"),
  rand = 123,
  tol = 1e-05,
  std_est = TRUE,
  formatted = FALSE,
  accuracy = 0.1
)
}
\arguments{
\item{x}{Numeric. The number of successes.}

\item{n}{Numeric. The number of trials.}

\item{conf_level}{Numeric. Confidence level for the interval, defaults to 0.95.}

\item{sides}{Character. Specifies the side of the confidence interval. Must be one of
\code{"two.sided"} (default), \code{"left"}, or \code{"right"}. Only the first value is used if multiple are provided.}

\item{method}{Character. The method used to compute the confidence interval. This can be one of
\code{"wilson"} (default), \code{"wald"}, \code{"waldcc"}, \code{"agresti-coull"}, \code{"jeffreys"}, \code{"modified wilson"},
\code{"wilsoncc"}, \code{"modified jeffreys"}, \code{"clopper-pearson"}, \code{"arcsine"}, \code{"logit"}, \code{"witting"},
\code{"pratt"}, \code{"midp"}, \code{"lik"}, or \code{"blaker"}. Abbreviations are accepted. Only the first value is used
if multiple are provided.}

\item{rand}{Numeric. Seed for the random number generator. Used for certain methods.}

\item{tol}{Numeric. Tolerance for the \code{"blaker"} method.}

\item{std_est}{Logical. If \code{TRUE} (default), returns the standard point estimator for the proportion
(x/n). If \code{FALSE}, returns the method-specific point estimate.}

\item{formatted}{Logical. If \code{TRUE}, formats the confidence intervals and point estimate as percentages.
Defaults to \code{FALSE}.}

\item{accuracy}{Numeric. The rounding accuracy for formatted percentages. Defaults to 0.1.}
}
\value{
A tibble with the following columns:
\itemize{
\item \code{percentage}: The point estimate of the proportion.
\item \code{lower_ci}: The lower bound of the confidence interval.
\item \code{upper_ci}: The upper bound of the confidence interval.
If \code{formatted = TRUE}, these values are formatted as percentages.
}
}
\description{
A wrapper around \code{DescTools::BinomCI} to calculate confidence intervals for binomial proportions.
This function formats the output into a tibble and optionally formats percentages.
}
\examples{
# Compute a Wilson confidence interval
binom_ci(x = 25, n = 100, method = "wilson")

# Compute a Clopper-Pearson confidence interval, formatted as percentages
binom_ci(x = 25, n = 100, method = "clopper-pearson", formatted = TRUE)

}
