% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blank_to_na.R
\name{blank_to_na}
\alias{blank_to_na}
\title{Replaces blank-ish elements of a factor or character vector to NA}
\usage{
blank_to_na(x, na_strings = c("", ".", "NA", "na", "N/A", "n/a", "NaN", "nan"))
}
\arguments{
\item{x}{a vector of factor or character or any type}

\item{na_strings}{case sensitive strings that will be coverted to NA. The
function will do a trimws(x,'both') before conversion. If NULL, do only
trimws, no conversion to NA.}
}
\value{
Returns a vector trimws (always for factor, character) and NA
converted (if matching na_strings). Attributes will also be kept
('label','labels', 'value.labels').
}
\description{
Replaces blank-ish elements of a factor or character vector to NA
}
\examples{
library(dplyr)
library(tibble)

blank_to_na(x = c("NA", "na", 1, 2, 3))

df <- tibble::tibble(
  a = c("NA", "na", 1, 2, 3),
  b = factor(c("NA", "na", 1, 2, 3))
)
df

levels(df$b)

df2 <- df \%>\%
  mutate(a = blank_to_na(a),
         b = blank_to_na(b))

df2
levels(df2$b)
}
\references{
https://stackoverflow.com/questions/24172111/change-the-blank-cells-to-na
}
