% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_2x2_res_table.R
\name{calc_2x2_res_table}
\alias{calc_2x2_res_table}
\title{Calculate agreement statistics for a 2 x 2 table}
\usage{
calc_2x2_res_table(table, has_gold_std = FALSE)
}
\arguments{
\item{table}{A table}

\item{has_gold_std}{Logical; \code{TRUE} will show the sensitivity and specificity
using rater __ as the gold standard. \code{FALSE} assumes there is no gold
standard and calculated the proportions of positive and negative agreement
instead.}
}
\value{
A tibble or data frame
}
\description{
I have opinions about what statistics should be reported along with kappa
when asssessing agreement between two raters. The chosen stats are based on
the advice in paper a paper by \strong{reference needed}.
}
\examples{
## Example 1 ----------------

foo <- matrix(c(20.61, 1.52, 9.58, 68.29), ncol = 2, byrow = TRUE)
foo <- as.table(100 * foo)
foo

calc_2x2_res_table(table = foo, has_gold_std = FALSE)
calc_2x2_res_table(table = foo, has_gold_std = TRUE)

## Example 2 ----------------

df <- tibble::tribble(
  ~a, ~b, ~c,
  1L, 1L, 1L,
  0L, 1L, 1L,
  1L, 1L, 0L,
  0L, 1L, 0L,
  1L, 1L, 1L,
  0L, 1L, 1L,
  1L, 1L, 0L
)

bar <- lamisc::make_table(df = df,
                          x_var = a,
                          x_lvls = c("1", "0"),
                          y_var = b,
                          y_lvls = c("1", "0"))

calc_2x2_res_table(table = bar, has_gold_std = FALSE)
calc_2x2_res_table(table = bar, has_gold_std = TRUE)
}
