% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_date_diff.R
\name{calc_date_diff}
\alias{calc_date_diff}
\title{calc_date_diff}
\usage{
calc_date_diff(dob, enddate = Sys.Date(), units = "months",
  precise = TRUE)
}
\arguments{
\item{precise}{}
}
\description{
Adapted from eeptools::calc_age
}
\examples{
a <- as.Date(seq(as.POSIXct('1987-05-29 018:07:00'), len = 26, by = "21 day"))
b <- as.Date(seq(as.POSIXct('2002-05-29 018:07:00'), len = 26, by = "21 day"))
calc_date_diff(a, units = 'years', precise = FALSE)

}
