% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_obs_agree.R
\name{calc_obs_agree}
\alias{calc_obs_agree}
\title{Calculate observed agreement and confidence interval}
\usage{
calc_obs_agree(data, ..., alpha = 0.05, tolerance = 0)
}
\arguments{
\item{data}{A data frame or tibble, n subjects (rows) and m raters (columns)}

\item{...}{Variables or columns of ratings}

\item{alpha}{Numeric; two-sided type one error, default = 0.05}

\item{tolerance}{Used in the \code{irr::agree()} function; number of successive
rating categories that should be regarded as rater agreement (see details).}
}
\value{
A tibble with the following columns
\item{N}{Number of subjects}
\item{po}{Percent agreement among raters}
\item{se_po}{Standard error of the proportion}
\item{conf_low}{Lower asymptotic confidence interval}
\item{conf_high}{Upper asymptotic confidence interval}
}
\description{
This is really just a wrapper for the \code{irr::agree()} function from that
package, but I wanted to calculate the confidence interval also.
}
\details{
Missing data are omitted in a listwise way.
Using extended percentage agreement (tolerance!=0) is only possible for
numerical values. If tolerance equals 1, for example, raters differing by one
scale degree are interpreted as agreeing.
}
\examples{
library(irr)
data("video")
calc_obs_agree(data = video, rater1:rater4)
calc_obs_agree(data = video, rater1:rater2)
}
