% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_or.R
\name{calc_or_wald}
\alias{calc_or_wald}
\title{Calculate Odds Ratio with Wald Confidence Intervals}
\usage{
calc_or_wald(n00, n01, n10, n11, alpha = 0.05)
}
\arguments{
\item{n00}{Number of cases where both variables are 0.}

\item{n01}{Number of cases where x = 0 and y = 1.}

\item{n10}{Number of cases where x = 1 and y = 0.}

\item{n11}{Number of cases where both variables are 1.}

\item{alpha}{Significance level for the two-sided Wald confidence interval (default is 0.05).}
}
\value{
A tibble containing odds ratio, lower confidence interval, upper confidence interval, and alpha value.
}
\description{
Computes the odds ratio between two binary variables along with the Wald confidence intervals.
}
