% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pw_kappa.R
\name{calc_pw_kappa}
\alias{calc_pw_kappa}
\title{Calculate pairwiese agreement statistics for multiple raters}
\usage{
calc_pw_kappa(data, ..., type = "unweighted")
}
\arguments{
\item{data}{A data frame or tibble}

\item{...}{Variable (column) names}

\item{type}{Character; "unweighted" or "weighted" kappa}
}
\value{
A list
}
\description{
When working with multiple-raters, it can be helpful to look at pairwise
agreement for all raters. The goal of this function is to automate some of
the steps incvolved to calculate statistics for each pair and summarize them
nicely in a table or a data frame.

Currently, pairwise kappa and proportion of obeserved agreement are the only
statistics available.
}
\examples{
diagnostic_df <- data.frame(stringsAsFactors = FALSE,
  id = c(1L, 2L, 3L, 4L, 5L, 6L, 7L, 8L, 9L, 10L, 11L, 12L, 13L, 14L,
         15L, 16L, 17L, 18L, 19L, 20L, 21L, 22L, 23L,
         24L, 25L, 26L, 27L, 28L, 29L, 30L),
  rater_1 = c("Yes", "No", "No", "No", "No", "No", "No", "No", "No", "No",
              "No", "No", "No", "No", "No", "No", "No",
              "No", "No", "No", "No", "Yes", "No", "No", "No",
              "No", "No", "No", "No", "No"),
  rater_2 = c("Yes", "No", "No", "No", "No", "No", "No", "No", "No", "No",
              "Yes", "No", "No", "Yes", "No", "No", "No",
              "No", "No", "No", "No", "Yes", "No", "No",
              "Yes", "No", "No", "No", "No", "No"),
  rater_3 = c("Yes", "No", "No", "No", "No", "No", "No", "No", "Yes", "No",
              "Yes", "Yes", "No", "Yes", "Yes", "No",
              "No", "No", "Yes", "No", "No", "Yes", "Yes",
              "Yes", "Yes", "No", "No", "Yes", "No", "No"),
  rater_4 = c("Yes", "No", "No", "No", "Yes", "No", "No", "No", "Yes", "No",
              "Yes", "Yes", "No", "Yes", "Yes", "No",
              "Yes", "No", "Yes", "No", "No", "Yes", "No",
              "Yes", "Yes", "No", "No", "Yes", "No", "No"),
  rater_5 = c("Yes", "No", "No", "No", "Yes", "No", "No", "No", "Yes", "No",
              "Yes", "Yes", "No", "Yes", "Yes", "No",
              "No", "No", "Yes", "No", "No", "Yes", "No", "Yes",
              "Yes", "No", "No", "Yes", "No", "No"),
  rater_6 = c("Yes", "No", "No", "No", "Yes", "No", "No", "Yes", "Yes",
              "No", "Yes", "Yes", "No", "Yes", "No", "No",
              "No", "No", "Yes", "No", "No", "No", "No",
              "Yes", "No", "Yes", "No", "Yes", "No", "No")
)


foo <- calc_pw_kappa(data = diagnostic_df,
                     rater_1:rater_6)
names(foo)
foo$k_table
foo$k_results
foo$k_min_max
foo$po_table
foo$po_results
foo$po_min_max

}
