% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_collin_diag.R
\name{calc_r_squared}
\alias{calc_r_squared}
\title{Calculate R-Squared Value}
\usage{
calc_r_squared(form, data)
}
\arguments{
\item{form}{A formula specifying the regression model (e.g., \code{"y ~ x1 + x2"}).}

\item{data}{A data frame containing the variables used in the formula.}
}
\value{
The R-squared value from the linear model.
}
\description{
This helper function computes the R-squared value for a given formula and dataset.
}
\examples{
data <- mtcars
form <- "mpg ~ disp + hp + wt"
calc_r_squared(form, data)

}
