% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sig_ind.R
\name{calc_sig_ind}
\alias{calc_sig_ind}
\title{Returns the significance codes the R shows in regression output.}
\usage{
calc_sig_ind(p_value, html = TRUE)
}
\arguments{
\item{p_value}{A p-value or vector of p-values. Can handle numeric or
characters with text.}

\item{html}{Logical; if \code{TRUE}, escapes character suitable for markdown,
else suitable for pdf or word.}
}
\value{
A character value or vector
}
\description{
Signif. codes:  \verb{0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1}
}
\examples{
library(dplyr)

pvals <- c(.133213, .06023, .004233, .000000134234, 1.0)
pvals_2 <- c(.133213, .06023, .004233, .000000134234, NA)
calc_sig_ind(pvals)
calc_sig_ind(pvals_2)

df <- tibble::tibble(p = pvals,
                     p_2 = pvals_2)

df \%>\%
  mutate(p_sig = calc_sig_ind(p),
         p_2_sig = calc_sig_ind(p_2)
  )

}
