% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_surv_table.R
\name{calc_surv_table}
\alias{calc_surv_table}
\title{Calculate time-to-event estimates}
\usage{
calc_surv_table(fit, all_group_name = "All subjects")
}
\arguments{
\item{fit}{\code{survfit} object}

\item{all_group_name}{String name for all records. Default is "All subjects".}
}
\value{
A tibble
}
\description{
Takes a \code{survfit} object and returns a tibble with median and confidence
intervals.
}
\examples{
library(survival)
library(dplyr)
head(aml)

fit <- survival::survfit(
  survival::Surv(time = time,
                 event = status) ~ x,
  data = aml)

calc_surv_table(fit)

fit <- survival::survfit(
  survival::Surv(time = time,
                 event = status) ~ 1,
  data = aml)

calc_surv_table(fit)
}
