% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ttest.R
\name{calc_ttest}
\alias{calc_ttest}
\title{Calculate t-test}
\usage{
calc_ttest(
  data,
  var,
  by = NULL,
  mu = 0,
  paired = TRUE,
  var_equal = FALSE,
  df_form = "Welch",
  conf_level = 0.95,
  check_variance = FALSE,
  reverse_groups = FALSE,
  show_cohens_d = FALSE
)
}
\arguments{
\item{data}{A tibble or data frame.}

\item{var}{A numeric vector of data values.}

\item{by}{Factor or character indicating groups}

\item{mu}{a number indicating the true value of the mean (or difference in means if you are performing a two sample test).}

\item{paired}{a logical indicating whether you want a paired t-test.}

\item{var_equal}{a logical variable indicating whether to treat the two variances as being equal. If TRUE then the pooled variance is used to estimate the variance otherwise the Welch (or Satterthwaite) approximation to the degrees of freedom is used.}

\item{df_form}{Character string indicating the formula to use to calculate the degrees of freedom when the variances are unequal: "Welch" (default) or "Satterthwaite"}

\item{conf_level}{confidence level of the interval.}

\item{check_variance}{If TRUE returns checks on the Hogeneity of variances. Default is FALSE}

\item{reverse_groups}{If TRUE, the \code{by} groups are reversed}

\item{show_cohens_d}{If TRUE, then estimated Cohen's d (effect size) is returned.}
}
\value{
A list
}
\description{
Perform two sample t-tests (one sample coming later). Wrapper around \code{t.test}
to return cleaner results and more information in one function.
}
\examples{

library(tibble)

fuel <- tibble::tribble(
  ~mpg, ~treated,
  20L,       0L,
  23L,       0L,
  21L,       0L,
  25L,       0L,
  18L,       0L,
  17L,       0L,
  18L,       0L,
  24L,       0L,
  20L,       0L,
  24L,       0L,
  23L,       0L,
  19L,       0L,
  24L,       1L,
  25L,       1L,
  21L,       1L,
  22L,       1L,
  23L,       1L,
  18L,       1L,
  17L,       1L,
  28L,       1L,
  24L,       1L,
  27L,       1L,
  21L,       1L,
  23L,       1L
)


# Unpaired, unequal variances, Welch DF
calc_ttest(data = fuel,
           var = mpg,
           by = treated,
           mu = 0,
           paired = FALSE,
           var_equal = FALSE,
           conf_level = 0.95)

# Unpaired, unequal variances, Satterthwaite DF
calc_ttest(data = fuel,
           var = mpg,
           by = treated,
           mu = 0,
           paired = FALSE,
           var_equal = FALSE,
           df_form = "Satterthwaite",
           conf_level = 0.95)


calc_ttest(data = fuel,
           var = mpg,
           by = treated,
           mu = 0,
           paired = FALSE,
           var_equal = TRUE,
           conf_level = 0.95)

calc_ttest(data = fuel,
           var = mpg,
           by = treated,
           mu = 0,
           paired = FALSE,
           var_equal = TRUE,
           conf_level = 0.95,
           show_cohens_d = TRUE)


calc_ttest(data = fuel,
           var = mpg,
           by = treated,
           mu = 0,
           paired = FALSE,
           var_equal = TRUE,
           conf_level = 0.95,
           reverse_groups = TRUE)

calc_ttest(data = fuel,
           var = mpg,
           by = treated,
           mu = 0,
           paired = FALSE,
           var_equal = FALSE,
           conf_level = 0.95,
           check_variance = TRUE)




### --------------------------------------------------------------
### Paired t-test, Flicker feather example, p. 185
### --------------------------------------------------------------

feather <- tibble::tribble(
  ~Bird,     ~name, ~value,
  "A", "Typical", -0.255,
  "A",     "Odd", -0.324,
  "B", "Typical", -0.213,
  "B",     "Odd", -0.185,
  "C", "Typical",  -0.19,
  "C",     "Odd", -0.299,
  "D", "Typical", -0.185,
  "D",     "Odd", -0.144,
  "E", "Typical", -0.045,
  "E",     "Odd", -0.027,
  "F", "Typical", -0.025,
  "F",     "Odd", -0.039,
  "G", "Typical", -0.015,
  "G",     "Odd", -0.264,
  "H", "Typical",  0.003,
  "H",     "Odd", -0.077,
  "I", "Typical",  0.015,
  "I",     "Odd", -0.017,
  "J", "Typical",   0.02,
  "J",     "Odd", -0.169,
  "K", "Typical",  0.023,
  "K",     "Odd", -0.096,
  "L", "Typical",   0.04,
  "L",     "Odd",  -0.33,
  "M", "Typical",   0.04,
  "M",     "Odd", -0.346,
  "N", "Typical",   0.05,
  "N",     "Odd", -0.191,
  "O", "Typical",  0.055,
  "O",     "Odd", -0.128,
  "P", "Typical",  0.058,
  "P",     "Odd", -0.182
)



calc_ttest(data = feather,
           var = value,
           by = name,
           paired = TRUE)
}
