% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_vif.R
\name{calc_vif}
\alias{calc_vif}
\title{Variance inflation factor (VIF)}
\usage{
calc_vif(fit)
}
\arguments{
\item{fit}{An \code{lm} or \code{glm} object}
}
\value{
A tbl_df
}
\description{
Calculates variance-inflation linear, generalized linear, and other models.

I found this function on
\href{https://stackoverflow.com/questions/33397689/multi-collinearity-for-categorical-variables}{StackOverflow}
The function \code{vif} from the \code{car} package would be my preference but I wanted to
put this one somewhere for posterity.
}
\examples{
library(car)

# Using the car package
car::vif(lm(prestige ~ income + education, data = Duncan))
car::vif(lm(prestige ~ income + education + type, data = Duncan))

# User defined function
calc_vif(lm(prestige ~ income + education, data = Duncan))
calc_vif(lm(prestige ~ income + education + type, data = Duncan))

# Another example using glm
model1 <- glm(case ~ spontaneous + induced,
              data = infert,
              family = binomial(link = "logit"))
summary(model1)
car::vif(model1)
calc_vif(model1)
}
