% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_spss_data.R
\name{clean_spss_data}
\alias{clean_spss_data}
\title{Clean SPSS Data Imported via \code{haven}}
\usage{
clean_spss_data(data, method = c("manual", "forcats"))
}
\arguments{
\item{data}{A data frame or tibble imported with \code{haven::read_sav()}.}

\item{method}{A string specifying how to convert labelled variables to factors.
Options are \code{"manual"} (uses \code{factor()} with \code{labels} attributes) or \code{"forcats"}
(uses \code{forcats::as_factor()} with \code{levels = "labels"}). Default is \code{"manual"}.}
}
\value{
A list with the following components:
\describe{
\item{data}{A cleaned data frame with labelled variables converted to factors.}
\item{dictionary}{A tibble showing variable names and labels.}
\item{factor_vars}{A named list of variables with labels, showing levels and names.}
}
}
\description{
This function cleans SPSS \code{.sav} data imported using \code{haven}, converting
labelled variables into factors and cleaning variable names. It also
creates a data dictionary and a list of factor variables with their levels
and labels.
}
\examples{
\dontrun{
# Raw_data can be downloaded from here and saved to disc
# https://www.cambridge.org/us/academic/subjects/psychology/psychology-research-methods-and-statistics/statistics-using-ibm-spss-integrative-approach-3rd-edition?format=PB&isbn=9781107461222
raw_data <- haven::read_spss(file = "/Users/latour/Dropbox/code/R/qualtrics/Wages.sav")
cleaned <- clean_spss_data(raw_data, method = "forcats")
str(cleaned$data)
}

}
