% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_all_char_to_fctr.R
\name{convert_all_char_to_fctr}
\alias{convert_all_char_to_fctr}
\title{Converts all character vectors to factors}
\usage{
convert_all_char_to_fctr(data)
}
\arguments{
\item{data}{Tibble or data frame}
}
\value{
A tibble
}
\description{
This function takes a tbl_df and returns same with any character column
converted to a factor.
}
\examples{
library(dplyr)
dplyr::glimpse(starwars)
convert_all_char_to_fctr(data = starwars)

}
