% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corandcov.R
\name{corandcov}
\alias{corandcov}
\title{Provides the Variance-covariance parameters estimates}
\usage{
corandcov(glsob, cov = TRUE, ...)
}
\arguments{
\item{glsob}{An object of class "gls" representing the linear model fit}

\item{cov}{Default is TRUE. Not sure how this is used}

\item{...}{Additional arguments}
}
\value{
A list with correlation matrix, variance structure, and covariance
matrix.
}
\description{
User-written function to output the covariance parameters
}
\examples{
library(dplyr)
library(readr)
library(nlme)

opposites <- readr::read_csv(
  file = "https://stats.idre.ucla.edu/stat/r/examples/alda/data/opposites_pp.txt",
  col_names = TRUE)
dplyr::glimpse(opposites)

unstruct <- gls(opp ~ time * ccog,
                opposites,
                correlation = corSymm(form = ~ 1 |
                                        id),
                weights = varIdent(form = ~ 1 | wave),
                method = "REML")

corandcov(unstruct)
}
\references{
\url{https://stats.idre.ucla.edu/r/examples/alda/r-applied-longitudinal-data-analysis-ch-7/}
}
