% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_chars.R
\name{count_chars}
\alias{count_chars}
\title{Character counting helper}
\usage{
count_chars(char)
}
\arguments{
\item{char}{a vector (atomic or list) or an expression object. Other objects
(including classed objects) will be coerced by base::as.list.}
}
\value{
A list
}
\description{
Function to help count / index the characters on the screen, rather than
using your fingers.
}
\examples{
count_chars(c("2020-12-26", "Klärschlammentsorgungsrichtlinie"))
}
\references{
https://rcrastinate.rbind.io/post/2020-12-26-little-helpers-character-index-counter/
}
