% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_count.R
\name{count_count}
\alias{count_count}
\title{Count and percentages}
\usage{
count_count(data, ..., sort = FALSE, .drop = FALSE)
}
\arguments{
\item{data}{A tibble or data frame.}

\item{...}{data-masking Variables to group by.}

\item{sort}{If TRUE, will show the largest groups at the top.}

\item{.drop}{For count(): if FALSE will include counts for empty groups (i.e. for levels of factors that don't exist in the data). Deprecated in add_count() since it didn't actually affect the output.}
}
\value{
An object of the same type as .data. count() and add_count() group transiently, so the output has the same groups as the input.
}
\description{
Wrapper function to \code{dplyr::count()} and \code{dplyr::add_count()}
}
\examples{
library(palmerpenguins)

count_count(data = penguins,
            species)

count_count(data = penguins,
            species,
            sex)

count_count(data = penguins,
            species,
            sex,
            sort = TRUE)
}
