% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_duplicates.R
\name{count_duplicates}
\alias{count_duplicates}
\title{Count the number of duplicate rows in a data frame}
\usage{
count_duplicates(data, ...)
}
\arguments{
\item{data}{A data frame or tibble}

\item{...}{Unquoted variable names to search for duplicates.}
}
\value{
Returns a data.frame (actually a \code{tbl_df}) with the full records where
the specified variables have duplicated values, as well as a variable
\code{dupe_count} showing the number of rows sharing that combination of
duplicated values.
}
\description{
Given a data frame, this will retun a data frame of the duplicate rows with a
column for the number of times that it appears in the data.

Very similar and not as preferred to the \code{get_dupes} function in Sam Firke's
\code{janitor} package. I did borrow some code from that one to deal with cases
when variable are specified and when they are not (variables are arguments to
\code{...}).
}
\examples{
library(dplyr)

(DF <- data.frame(replicate(sequence(1:3), n = 4)))
count_duplicates(DF)
count_duplicates(DF, X2, X3)
# Pipeable also
DF \%>\%
  count_duplicates(.)
}
\references{
\url{https://stackoverflow.com/questions/18201074/find-how-many-times-duplicated-rows-repeat-in-r-data-frame}

\url{https://cran.r-project.org/web/packages/janitor/janitor.pdf}

\url{https://github.com/sfirke/janitor/blob/master/R/get_dupes.R}
}
