% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts_to_cases.R
\name{counts_to_cases}
\alias{counts_to_cases}
\title{Convert from a table or data frame of counts to data frame of cases.}
\usage{
counts_to_cases(input_table, countcol = "Freq")
}
\arguments{
\item{input_table}{A table object}

\item{countcol}{the name of the column containing the counts
(default = "Freq")}
}
\value{
A data frame
}
\description{
Takes a data frame or a table object in R with counts and converts it to a
data frame with the cases. Source:
\url{http://www.cookbook-r.com/Manipulating_data/Converting_between_data_frames_and_contingency_tables/#countstocases-function}
}
\examples{
library(vcd)
class(SexualFun)
counts_to_cases(input_table = SexualFun,
                countcol = "Freq")

phs <- matrix(c(189, 10845, 104, 10933), byrow = TRUE, ncol = 2)
dimnames(phs) <- list(Group = c("Placebo", "Aspirin"), MI = c("Yes", "No"))
phs
class(phs)
counts_to_cases(input_table = phs,
                countcol = "Freq")
}
