% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_folder_name.R
\name{create_folder_name}
\alias{create_folder_name}
\title{Create a Clean Folder Name}
\usage{
create_folder_name(x, sep = "-")
}
\arguments{
\item{x}{A character string to be converted into a folder-friendly name.}

\item{sep}{A character used as the separator between words (default is "-").}
}
\value{
A character string formatted as a clean folder name.
}
\description{
This function takes a text string and returns a clean, standardized name for a folder.
}
\examples{
create_folder_name("My Folder Name") # Returns "my-folder-name"
create_folder_name("Another_example", sep = "_") # Returns "another_example"
}
