% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_proj_folders.R
\name{create_proj_folders}
\alias{create_proj_folders}
\title{Create folder with structure for analysis}
\usage{
create_proj_folders(main_dir = getwd(), sub_dir = NULL,
  folder_names = c("data", "code", "docs", "admin", "deliverables"))
}
\arguments{
\item{main_dir}{File path to the main directory where you want to create
folders. Default is the current working directory (\code{getwd()})}

\item{sub_dir}{Name of the sub-directory where you want to create folders.
Default is \code{Null}; folder can be completely specified using \code{main_dir} if
you'd like.}

\item{folder_names}{A character vector of the folder names to create in the
directory that you specified. Defaults are \code{c("data", "code", "docs", "admin", "deliverables")}.}
}
\description{
A helper function that will create folders for an analysis project in the
directory that is specified. Note that established folders should not be
overwritten since this funcion relies on \code{fs::dir_create()} but do use
caution.
}
\examples{
\dontrun{
create_proj_folders(main_dir = "/Users/latour/Desktop/",
                    sub_dir = "new_folder")
}
}
