% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimate_time.R
\name{decimate_time}
\alias{decimate_time}
\title{Convert time to decimals}
\usage{
decimate_time(time, units = "hours")
}
\arguments{
\item{time}{A date-time object or a character vector of hour minute second triples}

\item{units}{character string: "hours", "minutes", "seconds"}
}
\value{
A numeric
}
\description{
Given a class of "hms" or "difftime" or "character" ("00:00:00") or "Period", convert to decimal format.
}
\examples{
library(lubridate)

decimate_time(time = "3:30:00",
              units = "hours")

decimate_time(time = "3:30:00",
              units = "minutes")

decimate_time(time = lubridate::hms("3:30:00"),
              units = "minutes")

}
