% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop-fct-lvls.R
\name{drop_fct_lvls}
\alias{drop_fct_lvls}
\title{Drop factor levels and filter the data the same time}
\usage{
drop_fct_lvls(data, x, lvls)
}
\arguments{
\item{data}{A data frame or tibble}

\item{x}{A factor (or character vector).}

\item{lvls}{Character string of factors to drop}
}
\value{
An object of the same type as data.
}
\description{
When you filter data to remove rows that match a level in a factor, the
factor levels aren't removed. This function will filter the data and drop the
factor levels.
}
\examples{
library(dplyr)
library(palmerpenguins)

penguins \%>\%
  dplyr::count(species)

penguins \%>\%
  drop_fct_lvls(data = .,
                x = species,
                lvls = c("Chinstrap")) \%>\%
  dplyr::count(species)

penguins \%>\%
  drop_fct_lvls(data = .,
                x = species,
                lvls = c("Chinstrap",
                         "Gentoo")) \%>\%
  dplyr::count(species)
}
