% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_df_from_list.R
\name{extract_df_from_list}
\alias{extract_df_from_list}
\title{Take a list of data frames and write them to individual data frames with
names}
\usage{
extract_df_from_list(list_of_dfs, envir = .GlobalEnv)
}
\arguments{
\item{list_of_dfs}{A list with data frames}

\item{envir}{Environment to assign data frames to. Default is \code{.GlobalEnv}.}
}
\value{
Data frames; the same number as were contained in the list. Returned
to the global environment.
}
\description{
The function will take a list of data frames in the environment and extract
each data frame and assign the name from the list.
}
\examples{
library(readxl)
path <- readxl::readxl_example("datasets.xlsx")
list_of_sheets <- lamisc::read_excel_all_sheets(path_to_excel_file = path,
                                                as_list = TRUE)
extract_df_from_list(list_of_sheets)

}
\references{
\url{https://stackoverflow.com/questions/9726705/assign-multiple-objects-to-globalenv-from-within-a-function}
}
