% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_to_numeric.R
\name{factor_to_numeric}
\alias{factor_to_numeric}
\title{Convert a factor to integer/numeric}
\usage{
factor_to_numeric(x)
}
\arguments{
\item{x}{an object, for example a factor.}
}
\value{
A numeric object
}
\description{
Different than \code{lamisc::as_numeric_factor()}. This function returns
corresponding ordinal level value for the factor values. For a factor \code{f}, it
works like you would think that \code{as.numeric(f)} or
\code{as.numeric(as.character(f))} would work (both of these give an error). This
one is hard to explain but see the example.
}
\examples{
x1 <- c("Dec", "Apr", "Jan", "Mar")

month_levels <- c(
  "Jan", "Feb", "Mar", "Apr", "May", "Jun",
  "Jul", "Aug", "Sep", "Oct", "Nov", "Dec"
)

y1 <- factor(x1, levels = month_levels)
y1

factor_to_numeric(y1)

}
