% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fence.R
\name{fence}
\alias{fence}
\title{Limit values to a minimum and maximum}
\usage{
fence(x, lower = -Inf, upper = Inf, na_rm = FALSE)
}
\arguments{
\item{x}{Vector of numeric values.}

\item{lower}{Numeric. Lowest value.}

\item{upper}{Numeric. Highest value.}

\item{na_rm}{A logical indicating whether missing values should be removed.}
}
\value{
A numeric vector
}
\description{
Clip / clamp / or fence, i.e. limit, values to a lower and upper bound.
Similar to \code{raster::clamp()}. This is a wrapper around \code{base::pmin()} and
\code{base::pmax()}; see those for more details.
}
\examples{
#### Example 1 --------------------------------

vec <- sample(x = c(1:10),
       size = 10,
       replace = TRUE)

vec

fence(x = vec,
      lower = 4,
      upper = 6)

#### Example 2 --------------------------------

library(dplyr)
library(tibble)

df <- tibble::tibble(
  vec = vec
)

df \%>\%
  mutate(vec = fence(x = vec,
                     lower = 4,
                     upper = 6))
}
\references{
https://stackoverflow.com/questions/13868963/clip-values-between-a-minimum-and-maximum-allowed-value-in-r
}
