% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_func.R
\name{find_func}
\alias{find_func}
\title{Identifying the package of a function}
\usage{
find_func(f)
}
\arguments{
\item{f}{name of function for which the package(s) are to be identified; a
character string.}
}
\value{
A tibble.
}
\description{
I found this handy function to identify which package an R package belongs
to. I found this function on Sebastian Sauer's \href{https://sebastiansauer.github.io/finds_funs/}{blog}.
}
\examples{
\dontrun{
find_func("width")
}
}
