% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_messy_dates.R
\name{fix_messy_dates}
\alias{fix_messy_dates}
\title{Fix messy dates}
\usage{
fix_messy_dates(x, formats = c("m/d/y"), date_system = "modern")
}
\arguments{
\item{date_system}{}
}
\description{
Fix messy dates
}
\examples{
pacman::p_load(
tidyverse,    # packages ggplot2, tibble, tidyr, readr, purrr, and dplyr
janitor,      # for working with dirty data
lubridate,    # for working with dates and times
magrittr       # includes the \%<>\% assignment-pipe (\%>\% is loaded from dplyr)
)

bar <- tibble::tribble(
  ~date,                     ~comment,
  NA_character_,                      "",
  NA_character_,                      "",
  "12/21/2011",                       "",
  "2/1/2015",                         "",
  "5/17/12",                          "",
  "10/3/15",                          "",
  "42253",        "should be 2015-09-06",
  "42309",        "should be 2015-11-01",
  "5 /23/2015",                       "",
  "10/3 /2015",                       "",
  "unknown",    "Should turn into an NA",
  "10/5/10",      "should be Oct 5 2010",
  "2/1213",  "Can't handle this one yet",
  "21213",   "Can't handle this one yet"
)



bar \%<>\%
  mutate(date2 = fix_messy_dates(date)
  )

bar
}
