% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex_print.R
\name{flex_print}
\alias{flex_print}
\title{flextable wrapper}
\usage{
flex_print(
  x,
  font_name = "Arial",
  font_size = 11,
  align = NULL,
  align_j = NULL,
  auto_fit = TRUE,
  width = NULL,
  width_j = NULL,
  width_unit = "in",
  col_nms = NULL,
  col_lbls = NULL,
  title = NULL,
  subtitle = NULL,
  footer = NULL,
  title_size = 16,
  subtitle_size = 11,
  footer_size = 11
)
}
\arguments{
\item{x}{dataset}

\item{font_name}{single character value, the font family name. With Word and
PowerPoint output, the value specifies the font to be used to format
characters in the Unicode range (U+0000-U+007F). For "all" table parts.}

\item{font_size}{integer value (points). For "all" table parts.}

\item{align}{Column alignment: a character vector consisting of 'l' (left),
'c' (center), 'r' (right), and/or 'j' (justify). By default or if align =
NULL, numeric columns are right-aligned, and other columns are
left-aligned. If length(align) == 1L, the string will be expanded to a
vector of individual letters, e.g. 'clc' becomes c('c', 'l', 'c').}

\item{align_j}{Column selection for \code{align} argument. Default is
\code{1:length(align)}.}

\item{auto_fit}{Logical (Default is TRUE). See \code{flextable::autofit()}}

\item{width}{width in inches as a character vector.}

\item{width_j}{columns selection for width.}

\item{width_unit}{unit for width, one of "in", "cm", "mm".}

\item{col_nms}{Names of columns to be labelled.}

\item{col_lbls}{Labels to apply to columns different from column names.}

\item{title}{The text for the title.}

\item{subtitle}{The text for the subtitle which will be displayed below the
title.}

\item{footer}{The text for the footer which will be displayed below the body
of the table.}

\item{title_size}{Font size for the title. Default is 16.}

\item{subtitle_size}{Font size for the subtitle. Default is 11.}

\item{footer_size}{Font size for the footer Default is 11.}
}
\value{
A flextable
}
\description{
A wrapper function for \code{flextable::flextable()} with some of my preferred
formatting for simple tables.
}
\examples{
flex_print(x = head(iris))

flex_print(x = head(iris),
           font_name = "Lato",
           title = "Edgar Anderson's Iris Data",
           title_size = 16,
           subtitle = "Just the head")

flex_print(x = head(iris),
           font_name = "Lato",
           title = "Edgar Anderson's Iris Data",
           title_size = 16)

flex_print(x = head(iris),
           font_name = "Times New Roman",
           title = "Edgar Anderson's Iris Data",
           title_size = 16,
           footer = "The data were collected by Anderson, Edgar (1935). The irises of the Gaspe Peninsula, Bulletin of the American Iris Society, 59, 2–5.")

flex_print(x = head(iris),
           align = c("c"),
           col_lbls = c("Sepal Length",
                        "Sepal Width",
                        "Petal Length",
                        "Petal Width",
                        "Species"))


flex_print(x = head(iris),
           align = c("c"),
           col_nms = names(head(iris)),
           col_lbls = c("Sepal\nLength",
                        "Sepal\nWidth",
                        "Petal Length",
                        "Petal Width",
                        "Species"),
           font_size = 16)
}
