% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unique_col_vals.R
\name{get_unique_col_vals}
\alias{get_unique_col_vals}
\title{Extract the unique, non-duplicate values in a column}
\usage{
get_unique_col_vals(data, x)
}
\arguments{
\item{data}{A data frame or tibble}

\item{x}{Column in the data}
}
\value{
A character vector or \code{NULL}
}
\description{
Given a data frame and a column, the function will return the unique values
of that column. Useful when making factors.
}
\examples{
get_unique_col_vals(data = iris,
                    x = Species)

iris2 <- subset(iris, Species != "setosa")
table(iris2$Species)
get_unique_col_vals(data = iris2,
                    x = Species)

get_unique_col_vals(data = mtcars, x = gear)
}
