% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggladder.R
\name{ggladder}
\alias{ggladder}
\title{Ladder-of-powers histograms}
\usage{
ggladder(
  data,
  var,
  fill = "#0072B2",
  alpha = 0.8,
  line_color = "black",
  line_type = "solid"
)
}
\arguments{
\item{data}{A tibble or data frame.}

\item{var}{A numeric vector of data values.}

\item{fill}{Fill color for the histograms (default = "#0072B2").}

\item{alpha}{A numeric value (0–1) specifying the transparency of the histogram fill. Default is \code{0.8}.}

\item{line_color}{Color for the normal curves (default = "black").}

\item{line_type}{Line type for the normal curves (default = "solid").}
}
\value{
A tibble.
}
\description{
Explore Tukey's ladder of powers graphically using histograms to identify a
transform that converts \code{var}into a normally distributed variable.

Density is shown on the y-axis. The heights of the bars of the histogram
are scaled so that the sum of their areas is equal to 1.0. Normal curves are
plotted using the mean and standard deviation of the transformations, not the
kernel density of the histograms.

This is intended to replicate the \code{gladder} command in Stata. As per Stata's
approach, the number of bins is set to the \code{min(sqrt(n), 10 * log10(n))},
rounded to the closest integer, where n is number of observations. NAs
are omitted by default.
}
\examples{

ggladder(data = auto,
         var = mpg)

ggladder(data = auto,
         var = mpg,
         fill = "pink")

# To adjust the theme of all of the patches using patchwork package and an &
# instead of +
ggladder(data = auto,
         var = mpg) &
  ggplot2::theme_classic()

}
\references{
https://www.stata.com/manuals/rladder.pdf
}
