% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpreview.R
\name{ggpreview}
\alias{ggpreview}
\title{Preview a file that would be created by ggsave()}
\usage{
ggpreview(...)
}
\arguments{
\item{...}{Arguments passed to \code{ggplot2::ggsave()}}
}
\value{
An image in default system image viewer
}
\description{
Preview ggsave output in system image viewer. Credit for this one goes to TJ Mahr.
}
\examples{
\dontrun{
ggplot(mtcars) +
  aes(x = wt, y = mpg) +
  geom_point()

ggpreview(width = 3, height = 3, units = "in")

}
}
\references{
\url{https://gist.github.com/tjmahr/1dd36d78ecb3cff10baf01817a56e895}
}
