% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_in.R
\name{\%not_in\%}
\alias{\%not_in\%}
\title{Not in}
\usage{
x \%not_in\% table
}
\description{
Creating %not_in% with purrr::negate. Adapted from the cool, on going, Your
daily dose of {purrr} from Colin Fay.
\url{https://twitter.com/_ColinFay/status/987260548344631298/photo/1}
}
\details{
importFrom base \code{\%in}
}
\examples{
1 \%not_in\% 1:10
11 \%not_in\% 1:10
}
