% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hush.R
\name{hush}
\alias{hush}
\title{Sometimes you just need a little peace and quiet.}
\usage{
hush(code, is_unix = FALSE)
}
\arguments{
\item{code}{R code that you want to silence.}

\item{is_unix}{Logical; Default is FALSE, but if you are working in a UNIX
environment then you will want to set it to TRUE.}
}
\value{
The output of your code, but without printing to console.
}
\description{
This uses \code{sink()} to silence unwanted output from R code. So far it seems to
be more successful than \code{invisible()}. I basically followed the Stack
Overflow discussion in the references to create this one.
}
\examples{
\dontrun{
x <- dput(names(mtcars))
x <- hush(dput(names(mtcars)))
x
}
}
\references{
\url{https://stackoverflow.com/questions/2723034/suppress-one-commands-output-in-r}
}
