% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertBigBreakAddin.R
\name{insertBigBreakAddin}
\alias{insertBigBreakAddin}
\title{Insert big section break.}
\usage{
insertBigBreakAddin()
}
\description{
Call this function as an addin to insert section breaks that I used regularly
in my code. Meant to save some typing.
}
