% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertSmallBreakAddin.R
\name{insertSmallBreakAddin}
\alias{insertSmallBreakAddin}
\title{Insert small section break.}
\usage{
insertSmallBreakAddin()
}
\description{
Call this function as an addin to insert section breaks that I used regularly
in my code. Meant to save some typing.
}
