% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_kappa.R
\name{interpret_kappa}
\alias{interpret_kappa}
\title{Interpretation of kappa statistic}
\usage{
interpret_kappa(kappa)
}
\arguments{
\item{kappa}{Numeric; kappa statistic}
}
\value{
A string
}
\description{
This returns the interpretation of the strength of agreement determined by
the kappa statistic using the categorization by Landis and Koch (1977)
}
\examples{
interpret_kappa(0.5)
interpret_kappa(0.896)
interpret_kappa(-0.896)
interpret_kappa(0.60)

}
