% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_logit.R
\name{inv_logit}
\alias{inv_logit}
\title{Inverse logit function}
\usage{
inv_logit(logit)
}
\arguments{
\item{logit}{A numeric object. Missing values (NAs) are allowed.}
}
\value{
An object of the same type as \code{logit} containing the inverse logits of the input values.
}
\description{
Given a numeric object return the inverse logit of the values. Code adapted
from the \code{boot} package.
}
\details{
The inverse logit is defined by \code{exp(logit) / (1 + exp(logit))}. Values in \code{x} of \code{-Inf} or \code{Inf}
return logits of \code{0} or \code{1} respectively. Any \code{NA}s in the input will also be \code{NA}s
in the output.
}
\examples{
library(dplyr)
library(tibble)
library(titanic)
data(titanic_train, package = "titanic")
d <- titanic_train  # less typing
glm1 <- glm(Survived ~ Pclass, data = d, family = "binomial")
betas <- coef(glm1)

d <- d \%>\%
  tibble::as_tibble() \%>\%
  mutate(logit = betas[1] + betas[2] * Pclass,
         prob = inv_logit(logit = logit))

(preds <- predict(glm1, newdata = d, type = "response"))

d$prob == preds
}
