% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{la_palette}
\alias{la_palette}
\title{A few of my favorite and most used (discrete) palettes. Some are just base
R colors put into a vector. Others come from the follwoing sources:
https://github.com/mwaskom/seaborn/blob/master/seaborn/palettes.py
https://personal.sron.nl/~pault/}
\usage{
la_palette(name)
}
\arguments{
\item{name}{Name of desired palette. Choices are:
\code{basic_colors}, \code{qualitative_colors},  \code{tol_colors},
\code{sb_deep}, \code{sb_muted},  \code{sb_pastel}, \code{sb_bright},
\code{sb_dark},  \code{sb_colorblind}}
}
\value{
A vector of colours.
}
\description{
These are a handful of color palettes from Wes Anderson movies.
}
\examples{
library(lamisc)
la_palette("sb_deep")
la_palette("tol_colors")
la_palette("qualitative_colors")


}
\keyword{colors}
