% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{la_palettes}
\alias{la_palettes}
\title{Complete list of palettes}
\format{An object of class \code{list} of length 9.}
\usage{
la_palettes
}
\description{
Adapted from karthik/wesanderson
}
\keyword{datasets}
