% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ladder.R
\name{ladder}
\alias{ladder}
\title{Ladder of powers}
\usage{
ladder(data, var)
}
\arguments{
\item{data}{A tibble or data frame.}

\item{var}{A numeric vector of data values.}
}
\value{
A tibble.
}
\description{
Shows Tukey's ladder of powers to identify a transform that converts \code{var}
into a normally distributed variable.

This is intended to replicate the \code{ladder} command in Stata.
}
\examples{

ladder(data = auto,
       var = mpg)

}
\references{
https://www.stata.com/manuals/rladder.pdf
}
