% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lamisc.R
\docType{package}
\name{lamisc-package}
\alias{lamisc}
\alias{lamisc-package}
\title{lamisc: Miscellaneous helper functions}
\description{
The objective of the package is to be a landing spot for the miscellaneous
functions that I end up writing and want to preserve in one location.
}
\examples{
# Example usage:
library(lamisc)

}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/emilelatour/lamisc}
  \item Report bugs at \url{https://github.com/emilelatour/lamisc/issues}
}

}
\author{
\strong{Maintainer}: Emile Latour \email{latour@ohsu.edu}

}
\keyword{internal}
