% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_and_right.R
\name{left}
\alias{left}
\alias{right}
\title{Return text to the left and the right of a string}
\usage{
left(string, char)

right(string, char)
}
\arguments{
\item{string}{The string}

\item{char}{A character used to delimit the string}
}
\value{
A string
}
\description{
\code{left} takes a string and a character as arguments, then returns the
characters to the left. \code{right} takes a string and a character as
arguments, then returns the characters to the right.

Wrappers around some regular expressions. \code{Left} and \code{right} operate
similar to combining \code{left()}/\code{right()} and \code{search()} in excel. So maybe
there's a bit of nostalgia for me in using these functions. But I also
think that the naming makes it clearer what is going on.
}
\examples{
a <- c("left.right","left.right","left.right")
left(a, ".")
right(a, ".")

library(dplyr)
library(tibble)

test_df <- tibble::tibble(
  label = c("resection_margin=Mohs, > 1 cm",
            "resection_margin=Mohs, > 1 cm",
            "resection_margin=Mohs, <= 1 cm",
            "  resection_margin=Mohs, > 1 cm"))
test_df \%>\%
  mutate(left = left(label, "="),
         right = right(label, "="))

right(test_df$label, "=")
left(test_df$label, "=")
stringr::str_trim(left(test_df$label, "="), side = "both")

}
\references{
\url{https://stackoverflow.com/questions/45070117/r-check-if-special-character-in-string}

\url{https://stackoverflow.com/questions/36928870/r-check-if-string-contains-special-characters}
}
