% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/len_uniq.R
\name{len_uniq}
\alias{len_uniq}
\title{Number of unique values}
\usage{
len_uniq(vec, na.rm = TRUE)
}
\arguments{
\item{vec}{A vector}

\item{na.rm}{If \code{TRUE}, remove any missing values}
}
\value{
Number of unique values.
}
\description{
Get the number of unique values in a vector. Borrowed from Karl Borman.
Source: \url{https://github.com/kbroman/broman/blob/master/R/lenuniq.R}
}
\details{
It just does \code{length(unique(vec))} or, if
\code{na.rm=TRUE} (the default)
\code{length(unique(vec[!is.na(vec)]))}
}
\examples{
x <- c(1, 2, 1, 3, 1, 1, 2, 2, 3, NA, NA, 1)
len_uniq(x)
len_uniq(x, na.rm=FALSE)
}
\keyword{utilities}
