% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length2.R
\name{length2}
\alias{length2}
\title{Length that can handle NAs}
\usage{
length2(x, na_rm = FALSE)
}
\arguments{
\item{x}{A vector}

\item{na_rm}{Boolean; Default is \code{FALSE} which includes \code{NA}s. If \code{TRUE} \code{NA}s
will not be counted.}
}
\value{
an integer value
}
\description{
Version of length which can handle \code{NA}s: if \code{na.rm == TRUE}, don't count
them
}
\examples{
length2(x = c(1, 2, 3, NA, 5, 6, NA, 7, 8, NA, 10))
length2(x = c(1, 2, 3, NA, 5, 6, NA, 7, 8, NA, 10), na_rm = TRUE)
}
