% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lincomr.R
\name{lincomr}
\alias{lincomr}
\title{Calculate a linear combination from regression results.}
\usage{
lincomr(k, fit, alpha = 0.05)
}
\arguments{
\item{k}{A vector of 1's and 0's to indicate the linear combination}

\item{fit}{A model fit object}

\item{alpha}{Significance level}
}
\value{
A data frame of results
}
\description{
Pretty simple and strighforward right now. My friend Mike Lasarev had a big
hand in helping me get this set up and figuring out how to do it.

It should handle results from linear regression, mixed models, logistic
regression, and survival analysis. The big drawback currently is that
\emph{p}-values and test statistics are based on asymptotic theory; in the future
it would be a better idea to incorporate t-statistics or Satterthwaite
degrees of freedom.
}
\examples{
data(iris)
mod_1 <- lm(Petal.Width ~ Sepal.Length + Sepal.Width + Petal.Length,
            data = iris)
summary(mod_1)
lincomr(k = c(1, 0, 1, 1), fit = mod_1)
}
