% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dummies.R
\name{make_dummies}
\alias{make_dummies}
\title{Create Dummy Variables from Categorical Columns}
\usage{
make_dummies(data, vars)
}
\arguments{
\item{data}{A data frame containing the variables to be converted into dummy variables.}

\item{vars}{A character vector specifying the names of the categorical variables to be transformed into dummy variables.}
}
\value{
A data frame with the specified categorical variables replaced by their corresponding dummy variables.
}
\description{
The \code{make_dummies} function takes a data frame and a vector of variable names and returns a modified data frame where the specified categorical variables have been converted into dummy variables. The original variables are removed, and the dummy variables are added to the data frame with cleaned and consistent names.
}
\details{
This function uses \code{model.matrix} to create dummy variables for the specified categorical variables in the data. The dummy variables are created without an intercept (\code{- 1}) and are then cleaned using \code{janitor::clean_names} to ensure consistent naming. The function also uses \code{tibble::as_tibble} to return the dummy variables in tibble format. The original variables specified in \code{vars} are removed from the data frame and replaced with the newly created dummy variables.
}
\examples{
# Example usage
data <- data.frame(
  gender = c("Male", "Female", "Female", "Male"),
  age = c(25, 30, 22, 40)
)
make_dummies(data, vars = c("gender"))

}
