% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_palette.R
\name{my_palette}
\alias{my_palette}
\title{My preferred pallettes generator}
\usage{
my_palette(name, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of desired palette. Choices are:
\code{sb_deep}, \code{sb_muted},  \code{sb_pastel}, \code{sb_bright},
\code{sb_dark},  \code{sb_colorblind}, \code{lancet_colors},
\code{jco_colors}, \code{basic_colors}, \code{qualitative_colors},
\code{tol_colors}}

\item{type}{Either "continuous" or "discrete". Use continuous if you want
to automatically interpolate between colours.}
}
\value{
A vector of colours.
}
\description{
A few of my favorite and most used (discrete) palettes. Some are just base
R colors put into a vector. Others come from the follwoing sources:
https://github.com/mwaskom/seaborn/blob/master/seaborn/palettes.py
https://personal.sron.nl/~pault/
}
\details{
Adapted from \href{https://github.com/karthik/wesanderson}{karthik/wesanderson}
and \href{https://github.com/dill/beyonce}{dill/beyonce}
}
\examples{
library(lamisc)
library(tidyverse)
my_palette("tol_colors")
my_palette("tol_colors") \%>\%
  lamisc::view_palette()
ggplot(data = mtcars, aes(x = factor(cyl), fill = factor(vs))) +
  geom_bar(stat = "count") +
  scale_fill_manual(values = my_palette("tol_colors"))
ggplot(iris, aes(Sepal.Length, Sepal.Width, color = Species)) +
  geom_point(size = 3) +
  scale_color_manual(values = my_palette("lancet_colors")) +
  theme_minimal()


}
\keyword{colours}
