% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm-comb-my-sample.R
\name{my_sample}
\alias{my_sample}
\alias{perm}
\alias{comb}
\title{Counting, combinations, and permutations}
\usage{
my_sample(n, k, replace = TRUE, ordered = TRUE)

perm(n, k)

comb(n, k)
}
\arguments{
\item{n}{total number of objects in the set; an integer.}

\item{k}{number of choosing objects from the set; an integer}

\item{replace}{Logical indicating with or without replacement; default is
\code{TRUE}.}

\item{ordered}{Logical indicating ordered or unordered; default is
\code{TRUE}.}
}
\value{
A numeric value.
}
\description{
Functions to determine the number of possible ways to choose a sample
with/without replacement where order does and does not matter.

Without replacement, if the order doesn't matter then we have a combination,
if the order do matter then we have a permutation. One could say that a
permutation is an ordered combination.
}
\examples{
perm(15, 2)
comb(15, 2)

# same as comb(15, 2)
my_sample(n = 15,
          k = 2,
          replace = FALSE,
          ordered = FALSE)

# same as perm(15, 2)
my_sample(n = 15,
          k = 2,
          replace = FALSE,
          ordered = TRUE)

my_sample(n = 15,
          k = 2,
          replace = TRUE,
          ordered = TRUE)

my_sample(n = 15,
          k = 2,
          replace = TRUE,
          ordered = FALSE)
}
\references{
https://www.mathsisfun.com/combinatorics/combinations-permutations.html
}
