% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_preprocessing.R
\name{na_preprocessing}
\alias{na_preprocessing}
\title{NA preprocessing}
\usage{
na_preprocessing(dat)
}
\arguments{
\item{dat}{The full data set}
}
\value{
a data frame, with NA added as a level for factor / character.
}
\description{
Sometimes \code{na.omit} cannot be used with data set because it will end up with
too few rows to do anything sensible. This can be relaxed. In practice it is
the \code{NA_integer_} and \code{NA_real_} that have to be omitted, but \code{NA_character_} can
be retained: just add \code{NA} as a factor level. To achieve this, you need to loop
through variables in your data frame:
\itemize{
\item if a variable \code{x} is already a factor and \code{anyNA(x)} is \code{TRUE}, do \code{x <- addNA(x)}.
The "and" is important. If \code{x} has no \code{NA}, \code{addNA(x)} will add an unused \verb{<NA>}
level.
\item if a variable \code{x} is a character, do \code{x <- factor(x, exclude = NULL)} to
coerce it to a factor. \code{exclude = NULL} will retain \verb{<NA>} as a level.
\item if \code{x} is "logical", "numeric", "raw" or "complex", nothing should be
changed. \code{NA} is just \code{NA}.
}

\verb{<NA>} factor level will not be dropped by \code{droplevels} or \code{na.omit}, and it is
valid for building a model matrix. Check the following examples.

Once you add \code{NA} as a level in a factor / character, your dataset might
suddenly have more complete cases. Then you can run your model. If you still
get a "contrasts error", use \code{debug_contr_error2} to see what has happened.

Credit for this function goes to \href{https://stackoverflow.com/questions/44200195/how-to-debug-contrasts-can-be-applied-only-to-factors-with-2-or-more-levels-er}{this amazing stack overflow post}
}
\examples{
dat <- data.frame(y = 1:5,
                  x = factor(c(letters[1:4], NA)),
                  z = c(letters[1:4], NA))
dat
na_preprocessing(dat)

na.omit(dat)
na.omit(na_preprocessing(dat))
}
