% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_extract.R
\name{num_extract}
\alias{num_extract}
\title{Extraction of numbers from a character string}
\usage{
num_extract(string, as_char = TRUE)
}
\arguments{
\item{string}{A value or vector with mix of numbers and characters}

\item{as_char}{Logical; if \code{TRUE}, returns a character, and if
\code{FALSE} a numeric.}
}
\value{
A vector or value; character or numeric
}
\description{
Simple wrapper for some regex to extract only the numbers from a character
strin. Source: \url{http://stla.github.io/stlapblog/posts/Numextract.html}.
There are some other pretty cool functions after the link too.
}
\examples{
num_extract("30.5ml")

pvals <- c('0.00000', '< 0.001', '0.0', '0.123', '0.6', '1', '1.0', '1.000')
num_extract(pvals)
num_extract(pvals, as_char = FALSE)

negatives <- c(-0.133213, -0.06023, -0.004233, -0.000000134234, -1)
num_extract(negatives)
}
