% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pctr.R
\name{pctr}
\alias{pctr}
\title{Format proportion (0 < p < 1) to a percent (0 < pct < 100)}
\usage{
pctr(x, d = 0, as_text = TRUE, latex = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{d}{integer indicating the number of decimal places (round).}

\item{as_text}{logical; if FALSE, a numeric value is returned; if TRUE, then
a character is returned with a \% symbol}

\item{latex}{logical; default is \code{FALSE}. If \code{TRUE}, then the \% symbol is
escaped.}
}
\value{
An object of similar structure to \code{x}
}
\description{
Given a proportion (0 < p < 1), the function converts it to a percent (0 <
pct < 100) as either a numeric (when \code{as_text = FALSE}) or as a string with a
\% sign (\code{as_text = TRUE}). I owe this one to my buddy Kyle Hart.
\code{scales::percent()} has always been my preferred way to format percents but
this function has a little more flexibility and the number of decimal places
can be specified.
}
\examples{
pctr(x = runif(10), d = 1)
pctr(x = runif(10), d = 1, latex = TRUE)
pctr(x = runif(10), d = 1, as_text = FALSE)
}
