% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peep_dat.R
\name{peep_dat}
\alias{peep_dat}
\title{Peep at your data}
\usage{
peep_dat(data)
}
\arguments{
\item{data}{A data frame or tbl_df.}
}
\description{
Get a quick overview of your data set:
\itemize{
\item Number of observations,
\item Number of variables,
\item Number of duplicates,
\item Number of rows with missing values,
\item Number of columns with missing values,
\item Number of cells with missing values.
}

This is meant to just give some quick counts. I like to "peep" my data set
and paste this output in my code as I am cleaning and munging. For a deeper
look, I suggest \code{dplyr::glimpse()} which is where the inspiration for this
function except I don't always want to see the columns and I want more
details than \code{base::dim()}.

\code{janitor::get_dupes} is great for exploring duplicates. The \code{naniar} package
is best for exploring the missing values in your data.
}
\examples{
peep_dat(mtcars)

peep_dat(datasets::airquality)
}
