% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs
}
\description{
Like \code{dplyr}, \code{lamisc} also uses the pipe function, \code{\%>\%}
to turn function composition into a series of imperative statements.
}
\keyword{internal}
