% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_2x2_kable_table.R
\name{print_2x2_kable_table}
\alias{print_2x2_kable_table}
\title{Print a 2x2 table with kable}
\usage{
print_2x2_kable_table(data, x, y, x_name = NULL, y_name = NULL,
  format = "html")
}
\arguments{
\item{data}{A data frame or tibble}

\item{x}{"X" variable; counts appear in the rows of the table}

\item{y}{"Y" variable; counts appear in the columns of the table}

\item{x_name}{A string; name/label for the "X" variable}

\item{y_name}{A string; name/label for the "Y" variable}

\item{format}{A character string. Possible values are "latex", "html",
"markdown", "pandoc", and "rst". This is an argument for \code{knitr::kable()}}
}
\value{
A character vector of the table source code.
}
\description{
In my work with agreement statistics, I want to print a nice looking 2x2
table in the documents that I share with reaserchers. This function is a
wrapper for the steps that I take to make these types of 2x2 tables.
}
\examples{
library(dplyr)
library(ggplot2)
library(forcats)
library(janitor)

# Remove levels of some variables from the diamonds data frame so that it
# can be used as an example of a 2x2 table.
foo <- diamonds \%>\%
  dplyr::filter(color \%in\% c("E", "D")) \%>\%
  dplyr::filter(cut \%in\% c("Fair", "Good")) \%>\%
  dplyr::mutate(color = forcats::fct_drop(color),
                cut = forcats::fct_drop(cut))

# Tables using janitor::tabyl() and base::table()
foo \%>\% janitor::tabyl(color, cut)
table(foo$color, foo$cut)

# Example using the lamisc package
print_2x2_kable_table(data = foo,
                      x = color,
                      y = cut,
                      x_name = "COLOR",
                      y_name = "CUT")

}
