% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_pipe.R
\name{print_pipe}
\alias{print_pipe}
\title{Print result of a pipe}
\usage{
print_pipe(x, n = NULL)
}
\arguments{
\item{x}{An object or value from a previous pipe to print.}

\item{n}{Number of rows to print.}
}
\description{
Sometimes when I am working on a series of piped code, I want to see the
interim results when I am adding lines etc. It used to be that you could add
\verb{\%>\% \{.\}} with the pipe command from \code{magrittr} package (you still can), but
this does not work with the native pipe \verb{|>}. This is my workaround.
}
\details{
A wrapper around \code{x \%>\% print()} or \code{x |> print()}
}
\examples{
library(dplyr)
library(palmerpenguins)


penguins |>
  group_by(species) |>
  summarise(mean_bill_length_mm = mean(bill_length_mm, na.rm = TRUE)) |>
  print_pipe()

penguins |>
  print_pipe()

penguins |>
  print_pipe(n = Inf)
}
