% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_freq.R
\name{proc_freq}
\alias{proc_freq}
\title{A "PROC FREQ" for R}
\usage{
proc_freq(x, digits = 4)
}
\arguments{
\item{x}{A table or matrix}

\item{digits}{Doesn't do anything that I can tell}
}
\value{
A list
}
\description{
An R function to do some of the calcuations thatt SAS's PROC FREQ does. I
found this function on page 10 of \emph{An Introduction to Categorical Data
Analysis Using R} by Brett Presnell. It's just too fun not to include this
somehow. See references for link.
}
\examples{
jobsatis <- c(2,4,13,3, 2,6,22,4, 0,1,15,8, 0,3,13,8)
jobsatis <- matrix(jobsatis,byrow=TRUE,nrow=4)
dimnames(jobsatis) <- list(
  Income=c("<5","5-15","15-25",">25"),
  Satisfac=c("VD","LS","MS","VS"))
jobsatis
class(jobsatis)

jobsat_freq <- proc_freq(jobsatis)
jobsat_freq

library(vcd)
SexualFun
class(SexualFun)

proc_freq(SexualFun)

}
\references{
/url{http://users.stat.ufl.edu/~presnell/Courses/sta4504-2000sp/R/R-CDA.pdf}
}
