% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalr.R
\name{pvalr-deprecated}
\alias{pvalr-deprecated}
\title{A function for formatting p-values}
\usage{
pvalr(pvals, sig_limit, d, html)
}
\arguments{
\item{pvals}{A numeric value or vector of p-values}

\item{sig_limit}{Lower bound for precisions}

\item{d}{Integer indicating the number of decimal places (round)}

\item{html}{Logical; if \code{TRUE}, uses \code{&lt;} instead of \code{<}.}
}
\value{
A character value or vector of formatted p-values
}
\description{
Similar to the base R function \code{format.pval()} but does things the way that I
want them. Inspiration and lots of help with this one owed to
\url{https://github.com/raredd/rawr}.
}
\examples{
library(dplyr)

pvals <- c(.133213, .06023, .004233, .000000134234, 1.0)
pvals_2 <- c(.133213, .06023, .004233, .000000134234, NA)
pvalr(pvals)
pvalr(pvals_2)

df <- tibble::tibble(p = pvals,
                     p_2 = pvals_2)

df \%>\%
  mutate(p_formatted = pvalr(p),
         p_2_formatted = pvalr(p_2)
  )

}
\seealso{
\code{\link{lamisc-deprecated}}
}
\keyword{internal}
