% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqladder.R
\name{qqladder}
\alias{qqladder}
\title{Ladder-of-powers quantile-normal plots}
\usage{
qqladder(
  data,
  var,
  point_color = "cyan4",
  line_color = "black",
  line_type = "solid"
)
}
\arguments{
\item{data}{A tibble or data frame.}

\item{var}{A numeric vector of data values.}

\item{point_color}{Color for the quantile points (default = "cyan4").}

\item{line_color}{Color for the normal line (default = "black").}

\item{line_type}{Line type for the normal line (default = "solid").}
}
\value{
A tibble.
}
\description{
Explore Tukey's ladder of powers graphically using quantile plots to identify
a transform that converts \code{var}into a normally distributed variable.

This is intended to replicate the \code{qladder} command in Stata.
}
\examples{

qqladder(data = auto,
         var = mpg)

qqladder(data = auto,
         var = mpg,
         point_color = "purple")

qqladder(data = auto,
         var = weight,
         point_color = "purple",
         line_type = "dashed")


# To adjust the theme of all of the patches using patchwork package and an &
# instead of +
qqladder(data = auto,
         var = weight) &
  ggplot2::theme_classic()


}
\references{
https://www.stata.com/manuals/rladder.pdf
}
