% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel_all_sheets.R
\name{read_excel_all_sheets}
\alias{read_excel_all_sheets}
\title{Read in all the sheets at once from an excel file}
\usage{
read_excel_all_sheets(path_to_excel_file, as_list = TRUE)
}
\arguments{
\item{path_to_excel_file}{File path to the Excel file with multiple sheets}

\item{as_list}{Logical; if \code{TRUE} Returns a list of data frames. If \code{FALSE},
multiple data frames are returned to the global environment.}
}
\value{
A list of data frames. Or a number of data frames.
}
\description{
Given an excel file with muliple sheets, the function will iterate over each
sheet and save them all to a list or to individual data frames.
}
\examples{
\dontrun{
library(readxl)
path <- readxl::readxl_example("datasets.xlsx")

#### As a list --------------------------------

list_of_sheets <- read_excel_all_sheets(path_to_excel_file = path,
                                        as_list = TRUE)
list_of_sheets

#### As data frames --------------------------------

read_excel_all_sheets(path_to_excel_file = path,
                      as_list = FALSE)
}
}
\references{
\url{https://stackoverflow.com/questions/12945687/read-all-worksheets-in-an-excel-workbook-into-an-r-list-with-data-frames}

\url{https://readxl.tidyverse.org/articles/articles/readxl-workflows.html#iterate-over-multiple-worksheets-in-a-workbook}
}
