% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel_allsheets.R
\name{read_excel_allsheets}
\alias{read_excel_allsheets}
\title{Read in all sheets in an excel spreadsheet.}
\usage{
read_excel_allsheets(file, tibble = TRUE, ...)
}
\arguments{
\item{file}{Path to the xls/xlsx file}

\item{tibble}{\code{TRUE} to return a tibble, \code{FALSE} to return data frame.}

\item{...}{Additional arguments passes to \code{readxl::read_excel}}
}
\value{
A list of tibbles.
}
\description{
Helper function to quickly and easily read in all the sheets in an excel spreadsheet using \code{readxl::read_excel}
}
\examples{
library(readxl)
datasets <- readxl_example("datasets.xlsx")
read_excel_allsheets(datasets)
}
