% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_then_csv.R
\name{read_then_csv}
\alias{read_then_csv}
\alias{read_then_csv2}
\title{CSV caching and iterating over sheets}
\usage{
read_then_csv(sheet, path)

read_then_csv2(path)
}
\arguments{
\item{sheet}{String; Name of an excel sheet}

\item{path}{String; filepath to the the excel file}
}
\value{
\code{NULL}. CSV files are written to current working directory.
}
\description{
Read in all the sheets from an excel file at once and simultaneously cache to
CSV
}
\examples{
\dontrun{
library(dplyr)
library(readxl)
library(rlang)
library(purrr)

path <- readxl::readxl_example("datasets.xlsx")
path %>%
  readxl::excel_sheets() %>%
  rlang::set_names() %>%
  purrr::map(read_then_csv, path = path)

read_then_csv2(path)
}
}
\references{
\url{https://readxl.tidyverse.org/articles/articles/readxl-workflows.html#csv-caching-and-iterating-over-sheets}
}
