% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_obs_carried_fwd.R
\name{repeat_last}
\alias{repeat_last}
\title{A "Last Observation Carried Forward" function}
\usage{
repeat_last(x, forward = TRUE, maxgap = Inf, na.rm = FALSE)
}
\arguments{
\item{x}{A vector}

\item{forward}{Logical; default is \code{TRUE}. Direction to carry forward. By
specifying forward = FALSE, you can carry the last observation backward.}

\item{maxgap}{Numeric; By specifying maxgap, you can choose not to bridge
overly long gaps.}

\item{na.rm}{Logical; default is \code{FALSE}. Whether to remove NAs.}
}
\value{
A similar vector to x
}
\description{
Repeat the last non NA value. "Fill forward" NAs with the closest previous
non-NA value. There are lots of alternatives out there to this function. My
fave is \code{tidyr::fill()}.
}
\examples{
repeat_last(c(1,2,3,4,NA,NA))
repeat_last(c(1,NA,3,4,NA,NA), forward = FALSE)
repeat_last(c(1,NA,3,4,NA,NA, 5), forward = FALSE)

x = c(NA, NA, 1, NA, NA, NA, NA, NA, NA, 2, 3, 4, NA, NA, NA, NA, NA, 5, NA)
data.frame(x,
           repeat_last(x),
           repeat_last(x, forward = FALSE),
           repeat_last(x, maxgap = 5),
           check.names = FALSE)
}
\references{
\url{https://stackoverflow.com/questions/7735647/replacing-nas-with-latest-non-na-value}
}
