% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_lib_paths.R
\name{set_lib_paths}
\alias{set_lib_paths}
\title{Set your R library paths where you want when you want}
\usage{
set_lib_paths(lib_vec)
}
\arguments{
\item{lib_vec}{Character string with path to library}
}
\description{
Credit to this one goes to Miles McBain (see link in references). An
alternate to \code{.libPaths()} that will let you set your R library paths to
precisely whatever you choose, whenever you like.
}
\examples{
\dontrun{
.libPaths()
set_lib_paths("~/code/library")
.libPaths()
}
}
\references{
\href{https://milesmcbain.xyz/hacking-r-library-paths}{Hacking R's library paths}
}
