% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ttest.R
\name{skewness}
\alias{skewness}
\title{Internal function - calculates the skewness of a variable}
\usage{
skewness(x, na.rm = FALSE)
}
\arguments{
\item{x}{A (non-empty) numeric vector of data values.}
}
\description{
Internal function - calculates the skewness of a variable
}
\keyword{internal}
