% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split_2.R
\name{str_split_2}
\alias{str_split_2}
\title{strsplit 2.0 by Jakob Gepp}
\usage{
str_split_2(x, split, type = "remove", perl = FALSE, ...)
}
\arguments{
\item{x}{A character vector or string}

\item{split}{What to split the string by}

\item{type}{Select one of \code{type = c("remove", "before", "after")}}

\item{perl}{Extended results expresions are the default \code{perl = FALSE}
and \emph{Perl-like} regular expression with \code{perl = TRUE}}

\item{...}{Other arguments}
}
\value{
A list of characters
}
\description{
Splits the string but keeps the delimiter. Adopted from
\url{https://www.statworx.com/de/blog/strsplit-but-keeping-the-delimiter/}.
}
\details{
A regular expression is a pattern that describes a set of strings. Two types
of regular expressions are used in R, extended regular expressions (the
default) and Perl-like regular expressions used by perl = TRUE. There is a
also fixed = TRUE which can be considered to use a literal regular
expression.
}
\examples{
x <- c("3D/MON&SUN")
str_split_2(x, "(?<=.)(?=[/&])",perl = TRUE)

a <- c("asdasd.sss","segssddfge.sss","se.sss")
str_split_2(x = a, split = "[.]", type = "remove", perl = TRUE)

}
