% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_collin_diag.R
\name{syminv}
\alias{syminv}
\title{Symmetric Inverse}
\usage{
syminv(x)
}
\arguments{
\item{x}{A symmetric positive-definite matrix.}
}
\value{
The inverse of the input matrix.
}
\description{
Computes the inverse of a symmetric positive-definite matrix using its Cholesky decomposition.
}
\examples{
mat <- matrix(c(4, 2, 2, 3), ncol = 2)
syminv(mat)

}
