% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_palette.R
\name{view_palette}
\alias{view_palette}
\title{plot_colors and pie_palette}
\usage{
view_palette(palette)
}
\arguments{
\item{palette}{The palette that you want to visualize}
}
\description{
Various user defined palettes to use. And the function pie_palette() which
takes a palette as an argument and returns a pie chart with labels; handy
when you only want to pick specific colors from a palette and you don't know
what they are yet.
}
\examples{
library(lamisc)
library(beyonce)
view_palette(my_pals("sb_deep"))
view_palette(beyonce_palette(1))
}
