% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_multi_tabbed_xlsx.R
\name{write_multi_tabbed_xlsx}
\alias{write_multi_tabbed_xlsx}
\title{Easily Make Multi-tabbed .xlsx Files with openxlsx}
\usage{
write_multi_tabbed_xlsx(tab_names, list_of_dfs, file_path, over_write = TRUE)
}
\arguments{
\item{tab_names}{Character vector of names for the tabs in the Excel
spreadsheet.}

\item{list_of_dfs}{A list of data frames to write to file.}

\item{file_path}{Path and filename to write to.}

\item{over_write}{Logical. Overwrite the existing file of the same name.
Default is \code{TRUE}}
}
\value{
A workbook object
}
\description{
Wrapper for some steps using \code{openxlsx} package to write multiple to multiple
tabs in an Excel spreadsheet. Apparently, \code{openxlsx::write.xlsx()} will
accomplish the same thing. I just found this out. I use this other method so
regularly that I still feel like saving it here for posterity.
}
\examples{
\dontrun{
dat <- split(mtcars, mtcars$cyl)
dat

write_multi_tabbed_xlsx(tab_names = c("4 cyl", "6 cyl", "8 cyl"),
                        list_of_dfs = dat,
                        file_path = "C:/Users/latour/Desktop/cylinders.xlsx",
                        over_write = TRUE)
}
}
\references{
\itemize{
\item \href{https://trinkerrstuff.wordpress.com/2018/02/14/easily-make-multi-tabbed-xlsx-files-with-openxlsx/}{TRinker's R Blog}

}
}
