% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tabyl.R
\name{calc_tabyl}
\alias{calc_tabyl}
\title{Make a tabyl with totals and formatting}
\usage{
calc_tabyl(df, x)
}
\arguments{
\item{df}{A data frame or tibble.}

\item{x}{A single variable that you want to see the formatted tabyl for.}
}
\description{
This is a simple wrapper around the tabyl function from the janitor
package. So rather than writing a bunch of lines over and over again, I
made this shortened function to make my life easier. Since the first
argument takes a data frame or a tibble, this function is pipeable.
}
\examples{
iris \%>\%
  calc_tabyl(Species)

calc_tabyl(df = iris, x = Species)

}
