% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{fmt_num}
\alias{fmt_num}
\alias{fmt_pct}
\title{Number formatters}
\usage{
fmt_num(
  x,
  accuracy = 1,
  scale = 1,
  prefix = "",
  suffix = "",
  big.mark = "",
  decimal.mark = ".",
  trim = TRUE,
  as_numeric = FALSE,
  ...
)

fmt_pct(
  x,
  accuracy = NULL,
  scale = 100,
  prefix = "",
  suffix = "\%",
  big.mark = "",
  decimal.mark = ".",
  trim = TRUE,
  ...
)
}
\arguments{
\item{x}{A numeric vector to format.}

\item{accuracy}{Number to round to, \code{NULL} for automatic guess.}

\item{scale}{A scaling factor: \code{x} will be multiply by \code{scale} before
formating (useful if the underlying data is on another scale,
e.g. for computing percentages or thousands).}

\item{prefix, suffix}{Symbols to display before and after value.}

\item{big.mark}{Character used between every 3 digits to separate thousands.}

\item{decimal.mark}{The character to be used to indicate the numeric
decimal point.}

\item{trim}{Logical, if \code{FALSE}, values are right-justified to a common
width (see \code{\link[base:format]{base::format()}}).}

\item{as_numeric}{Logical; if TRUE, a numeric value is returned}

\item{...}{Other arguments passed on to \code{\link[base:format]{base::format()}}.}
}
\value{
\verb{*_format()} returns a function with single parameter
\code{x}, a numeric vector, that returns a character vector.
}
\description{
A set of functions to format numeric values:
\itemize{
\item \code{fmt_num()} is a generic formatter for numbers.
\item \code{fmt_pct()} multiply values by one hundred and display percent sign.
}

All formatters allow you to re-\code{scale} (multiplicatively), to round to
specified \code{accuracy}, to add custom \code{suffix} and \code{prefix} and to specify
\code{decimal.mark} and \code{big.mark}.

The functions were adapted directly from the \code{scales} package.
}
\examples{
v <- c(12.3, 4, 12345.789, 0.0002)
fmt_num(v)
fmt_num(v, big.mark = ",")
fmt_num(v, accuracy = .001)
fmt_num(v, accuracy = .001, decimal.mark = ",")
fmt_num(v, accuracy = .5)

w <- c(-90.256, -0.1212, NA, NA)
fmt_num(w)
fmt_num(w, big.mark = ",")
fmt_num(w, accuracy = .001)
fmt_num(w, accuracy = .001, decimal.mark = ",")
fmt_num(w, accuracy = .5)
fmt_pct(x = runif(10))
fmt_pct(x = c(1, runif(10)), trim = FALSE)
}
