% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_print.R
\name{k_print}
\alias{k_print}
\title{Kable print wrapper}
\usage{
k_print(
  x,
  format,
  col_names = NA,
  bootstrap_options = c("striped"),
  escape = TRUE,
  align = NULL,
  long_table = FALSE,
  latex_options = "basic",
  caption = NULL,
  cols = NULL,
  widths = NULL
)
}
\arguments{
\item{x}{An R object, typically a matrix or data frame.}

\item{format}{A character string. Possible values are latex, html, markdown,
pandoc, and rst; this will be automatically determined if the function is
called within knitr; it can also be set in the global option
knitr.table.format. If format is a function, it must return a character
string.}

\item{col_names}{A character vector of column names to be used in the table.}

\item{bootstrap_options}{A character vector for bootstrap table options.
Please see package vignette or visit the w3schools' Bootstrap Page for more
information. Possible options include basic, striped, bordered, hover,
condensed and responsive.}

\item{escape}{Boolean; whether to escape special characters when producing
HTML or LaTeX tables. When escape = FALSE, you have to make sure that
special characters will not trigger syntax errors in LaTeX or HTML.}

\item{align}{Column alignment: a character vector consisting of 'l' (left),
'c' (center) and/or 'r' (right). By default or if align = NULL, numeric
columns are right-aligned, and other columns are left-aligned. If
length(align) == 1L, the string will be expanded to a vector of individual
letters, e.g. 'clc' becomes c('c', 'l', 'c'), unless the output format is
LaTeX.}

\item{long_table}{Boolean; Whether to break a table across multiple pages or
not.}

\item{latex_options}{A character vector for LaTeX table options. Please see
package vignette for more information. Possible options include basic,
striped, hold_position, HOLD_position, scale_down & repeat_header. striped
will add alternative row colors to the table. It will imports LaTeX package
xcolor if enabled. hold_position will "hold" the floating table to the
exact position. It is useful when the LaTeX table is contained in a table
environment after you specified captions in \code{kable()}. It will force the
table to stay in the position where it was created in the document. A
stronger version: HOLD_position requires the float package and specifies.
scale_down is useful for super wide table. It will automatically
adjust the table to page width. repeat_header in only meaningful in a
longtable environment. It will let the header row repeat on every page in
that long table.}

\item{caption}{The table caption.}

\item{cols}{A numeric value or vector indicating which column(s) to be selected.}

\item{widths}{A character string telling HTML & LaTeX how wide the column
needs to be, e.g. "10cm", "3in" or "30em".}
}
\value{
A character vector of the table source code.
}
\description{
A wrapper function for \code{knitr::kable} and \code{kableExtra} package.
}
\examples{
\dontrun{
k_print(x = head(iris),
        align = c("c", "c", "c", "c", "r"),
        format = "html",
        cols = c(2, 3, 4),
        widths = c("10em", "20em", "10em"))
        }
}
