% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left-and-right.R
\name{left}
\alias{left}
\alias{right}
\title{Return text to the left and the right of a string}
\usage{
left(string, char)

right(string, char)
}
\arguments{
\item{string}{The string}

\item{char}{A character used to delimit the string}
}
\value{
A string
}
\description{
\code{left} takes a string and a character as arguments, then returns the
characters to the left. \code{right} takes a string and a character as
arguments, then returns the characters to the right.
}
\details{
Wrappers around some regular expressions. \code{Left} and \code{right} operate
similar to combining \code{left()}/\code{right()} and \code{search()} in excel. So maybe
there's a bit of nostalgia for me in using these functions. But I also
think that the naming makes it clearer what is going on.

A couple of reference:
\itemize{
\item \url{https://stackoverflow.com/questions/45070117/r-check-if-special-character-in-string}
\item \url{https://stackoverflow.com/questions/36928870/r-check-if-string-contains-special-characters}
}
}
\examples{
a <- c("asdasd.sss","segssddfge.sss","se.sss")
left(a, ".")
right(a, ".")


}
