% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggladder.R
\name{make_histos}
\alias{make_histos}
\title{Internal function - Make the histograms}
\usage{
make_histos(
  df,
  x,
  x_title,
  fill = "cyan4",
  line_color = "black",
  line_type = "solid"
)
}
\arguments{
\item{df}{A data frame or tibble.}

\item{x}{A (non-empty) numeric vector of data values.}

\item{x_title}{The title for the plot}

\item{fill}{Color for the histograms}

\item{line_color}{Color for the line}

\item{line_type}{Line type}
}
\description{
Internal function - Make the histograms
}
\keyword{internal}
