% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggladder.R
\name{make_histos}
\alias{make_histos}
\title{Internal function - Make the histograms}
\arguments{
\item{df}{A data frame or tibble.}

\item{x}{A (non-empty) numeric vector of data values.}

\item{x_title}{The title for the plot}

\item{fill}{Color for the histograms}

\item{line_color}{Color for the line}

\item{line_type}{Line type}
}
\description{
This function generates histograms for a given numeric vector in a data frame.
}
\keyword{internal}
