% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_quarto_template.R
\name{use_quarto_template}
\alias{use_quarto_template}
\title{Create a quarto template}
\usage{
use_quarto_template(file_name = NULL)
}
\arguments{
\item{file_name}{name of the sub-directory and new qmd report within it}
}
\description{
I followed \href{https://spencerschien.info/post/r_for_nonprofits/quarto_template/}{this great post} to
translate R Markdown’s templating within R packages to the Quarto protocol. I
wanted to be able to easily create the skeleton for my preferred quarto
settings without copying and pasting them from a file over and over again.

All of the extension stuff, I don' care about and it has been commented out
}
